Port to TypeScript, add fpmount functionality. (#2)
* Integrate fpmount, port to typescript, and copy Colin's extension folder layout.

* Forgot slash before mount.php for docker case.

* Optimize imports.

* Ignore games with the '-extract' parameter.

* Use npm typings, instead of local ones.

* Moved to promises, added nudge, and commented.

The promises ensure that AutoMount will block until the mounting is
complete. This should (in theory) prevent games from getting 404s
because the mounting is too slow.

The nudges: add some complicated timer logic, and maybe all your
problems will be solved. Current nudge-timeout is set at two seconds.

Comments: I added some, because I didn't want to force a reader
to try to understand the monstrosity that this is becoming.